debug

+++ src/main/java/cn/anytec/ffmpeg/FFMPEGService.java
@@ -9,7 +9,7 @@ public interface FFMPEGService {
 
     boolean cutVideo(File source, File output, String start, String duration, boolean isCover);
 
-    boolean deferVideo(File source, File output, Integer fps, boolean isCover);
+    boolean deferVideo(File source, File output, double fps, boolean isCover);
 
     boolean concatMedia(File concatFile, File output, boolean isCover);
